+++ tests/test_db.py
@@ -1,5 +1,6 @@
 import unittest
-from ..ts_process.database import DataBase
+
+from ts_process import database as db
 
 class TestDataBase(unittest.TestCase):
     def setUp(self):
@@ -7,16 +8,16 @@ class TestDataBase(unittest.TestCase):
         self.csize = 10
 
     def test_create_database_singleton(self):
-        db_1 = DataBase(self.dbname, self.csize)
-        db_2 = DataBase(self.dbname, self.csize)
+        db_1 = db.DataBase(self.dbname, self.csize)
+        db_2 = db.DataBase(self.dbname, self.csize)
 
         self.assertEqual(db_1,db_2)
-
-
-
-
-if __name__ == "__main__":
-    unittest.main()
+    
+    def test_set_and_get_value(self):
+        db_3 = db.DataBase('mytest',10)
+        db_3.set_value('x1',100)
+        self.assertEqual(db_3.get_value('x1'),100)
+