+++ CHANGELOG.md
@@ -6,6 +6,8 @@ See [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) for more details.
 ### Added
 - database.py module and DataBase class
 - MANIFEST.in file
+- tests folder
+- test_db.py file
 
 ### Changed